Add authentication for when a user wishes to edit and cancel a booking and also add the test to check that the test asserts a 302 code when the user cancels a booking 302 = redirect

+++ Accounts/views.py
@@ -6,11 +6,13 @@ from Staff.forms import EditBookingForm
 # Create your views here.
 
 def show_user_reservations(request):
-    booking = Booking.objects.filter(user=request.user)
-    context = {
-        "booking": booking
-    }
-    return render(request, "user_reservations.html", context)
+    if request.user.is_authenticated:
+        booking = Booking.objects.filter(user=request.user)
+        context = {
+            "booking": booking
+        }
+        return render(request, "user_reservations.html", context)
+    return HttpResponseRedirect("/")
 
 
 def user_details_booking(request, booking_id):
@@ -22,31 +24,36 @@ def user_details_booking(request, booking_id):
 
 
 def user_edit_booking(request, booking_id):
-    booking_data = get_object_or_404(Booking, id=booking_id)
-    booking = Booking.objects.filter(id=booking_id)
-    next = request.POST.get("next", "/")
-    if request.method == "POST":
-        form = EditBookingForm(request.POST, instance=booking_data)
-        if form.is_valid():
-            instance = form.save(commit=False)
-            for item in booking:
-                item.booking_approved = False
-                item.booking_acknowledged = False
-                item.save()
-            instance.save()
-            success_context = {
-                "data": instance,
-            }
-            return render(request, "user_edit_booking_success.html", success_context)
-    form = EditBookingForm(instance=booking_data)
-    context = {
-        "booking": booking,
-        "form": form
-    }
-    return render(request, "user_edit_booking.html", context)
+    if request.user.is_authenticated:
+        booking_data = get_object_or_404(Booking, id=booking_id)
+        booking = Booking.objects.filter(id=booking_id)
+        if request.method == "POST":
+            form = EditBookingForm(request.POST, instance=booking_data)
+            if form.is_valid():
+                instance = form.save(commit=False)
+                for item in booking:
+                    item.booking_approved = False
+                    item.booking_acknowledged = False
+                    item.save()
+                instance.save()
+                success_context = {
+                    "data": instance,
+                }
+                return render(request, "user_edit_booking_success.html", success_context)
+        form = EditBookingForm(instance=booking_data)
+        context = {
+            "booking": booking,
+            "form": form
+        }
+        return render(request, "user_edit_booking.html", context)
+    else:
+        return HttpResponseRedirect("/")
 
 
 def user_cancel_booking(request, booking_id):
-    booking = get_object_or_404(Booking, id=booking_id)
-    booking.delete()
-    return redirect("user_reservations")
+    if request.user.is_authenticated:
+        booking = get_object_or_404(Booking, id=booking_id)
+        booking.delete()
+        return redirect("user_reservations")
+    else:
+        return HttpResponseRedirect("/")
